reformatted level data to prevent UTF-16 encoding. fixed weird collision bug on large levels. tweaked border portals

+++ dependencies/sfml/extlibs/headers/stb_image/stb_image.h
@@ -535,7 +535,7 @@ typedef unsigned int   stbi__uint32;
 typedef   signed int   stbi__int32;
 #else
 #include <stdint.h>
-typedef uint16_t stbi__uint16;
+typedef uint32_t stbi__uint16;
 typedef int16_t  stbi__int16;
 typedef uint32_t stbi__uint32;
 typedef int32_t  stbi__int32;
