+++ src/entities/world/Portal.hpp
@@ -10,14 +10,14 @@
 
 namespace entity {
 
-	const uint16_t CELL_SIZE = 32;
+	const uint32_t CELL_SIZE = 32;
 
 	class Portal {
 
 	public:
 
 		using Vec = sf::Vector2<float>;
-		using Vecu16 = sf::Vector2<uint16_t>;
+		using Vecu16 = sf::Vector2<uint32_t>;
 
 		Portal() = default;
 		Portal(Vecu16 dim, Vecu16 pos) : scaled_dimensions(dim), scaled_position(pos) {
